cleaning/curse_dimensionality_sim/requirements.txt
torch>=1.9.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
tqdm>=4.61.0
pyyaml>=5.4.0